% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifGlmnet.R
\docType{data}
\name{LearnerClassifGlmnet}
\alias{LearnerClassifGlmnet}
\alias{mlr_learners_classif.glmnet}
\title{GLM with Elastic Net Regularization Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\description{
Generalized linear models with elastic net regularization.
Calls \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} from package \CRANpkg{glmnet}.
\insertNoCite{friedman_2010}{mlr3learners}
}
\section{Construction}{
\preformatted{LearnerClassifGlmnet$new()
mlr3::mlr_learners$get("classif.glmnet")
mlr3::lrn("classif.glmnet")
}
}

\examples{
learner = mlr3::lrn("classif.glmnet")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
\insertAllCited{}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
