% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifKKNN.R
\docType{data}
\name{LearnerClassifKKNN}
\alias{LearnerClassifKKNN}
\alias{mlr_learners_classif.kknn}
\title{k-Nearest-Neighbor Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifKKNN
}
\description{
k-Nearest-Neighbor classification.
Calls \code{\link[kknn:kknn]{kknn::kknn()}} from package \CRANpkg{kknn}.
}
\keyword{datasets}
