% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifLDA.R
\name{LearnerClassifLDA}
\alias{LearnerClassifLDA}
\alias{mlr_learners_classif.lda}
\title{Linear Discriminant Analysis Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\description{
Linear discriminant analysis.
Calls \code{\link[MASS:lda]{MASS::lda()}} from package \CRANpkg{MASS}.
}
\section{Construction}{
\preformatted{LearnerClassifLDA$new()
mlr3::mlr_learners$get("classif.lda")
mlr3::lrn("classif.lda")
}
}

\examples{
learner = mlr3::lrn("classif.lda")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
\cite{mlr3learners}{venables_2002}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
