% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifQDA.R
\docType{data}
\name{LearnerClassifQDA}
\alias{LearnerClassifQDA}
\alias{mlr_learners_classif.qda}
\title{Quadratic Discriminant Analysis Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\description{
Quadratic discriminant analysis.
Calls \code{\link[MASS:qda]{MASS::qda()}} from package \CRANpkg{MASS}.

\insertNoCite{venables_2002}{mlr3learners}
}
\section{Construction}{
\preformatted{LearnerClassifQDA$new()
mlr3::mlr_learners$get("classif.qda")
mlr3::lrn("classif.qda")
}
}

\examples{
learner = mlr3::lrn("classif.qda")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
\insertAllCited{}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
