% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrLM.R
\name{LearnerRegrLM}
\alias{LearnerRegrLM}
\alias{mlr_learners_regr.lm}
\title{Linear Model Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\description{
Ordinary linear regression. Calls \code{\link[stats:lm]{stats::lm()}}.

The importance scores are calculated using \code{-log10(p)} where \code{p} is the \eqn{p}-value.
This transformation is necessary to ensure numerical stability for very small \eqn{p}-values.
}
\section{Construction}{
\preformatted{LearnerRegrLM$new()
mlr3::mlr_learners$get("regr.lm")
mlr3::lrn("regr.lm")
}
}

\examples{
library(mlr3)
learner = lrn("regr.lm")
print(learner)

# available parameters:
learner$param_set$ids()

task = tsk("mtcars")
learner$train(task)
summary(learner$model)

pred = learner$predict(task, row_ids = head(task$row_ids))
pred
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
