% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrRanger.R
\name{LearnerRegrRanger}
\alias{LearnerRegrRanger}
\alias{mlr_learners_regr.ranger}
\title{Ranger Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\description{
Random regression forest.
Calls \code{\link[ranger:ranger]{ranger::ranger()}} from package \CRANpkg{ranger}.
}
\section{Construction}{
\preformatted{LearnerRegrRanger$new()
mlr3::mlr_learners$get("regr.ranger")
mlr3::lrn("regr.ranger")
}
}

\examples{
learner = mlr3::lrn("regr.ranger")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
\cite{mlr3learners}{wright_2017}

\cite{mlr3learners}{breiman_2001}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
