% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrSVM.R
\name{LearnerRegrSVM}
\alias{LearnerRegrSVM}
\alias{mlr_learners_regr.svm}
\title{Support Vector Machine}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\description{
A learner for a regression support vector machine implemented in \code{\link[e1071:svm]{e1071::svm()}}.
}
\section{Construction}{
\preformatted{LearnerRegrSVM$new()
mlr3::mlr_learners$get("regr.svm")
mlr3::lrn("regr.svm")
}
}

\examples{
learner = mlr3::lrn("regr.svm")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
\cite{mlr3learners}{cortes_1995}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
