% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrXgboost.R
\docType{data}
\name{LearnerRegrXgboost}
\alias{LearnerRegrXgboost}
\alias{mlr_learners_regr.xgboost}
\title{Extreme Gradient Boosting Regressiong Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrXgboost
}
\description{
eXtreme Gradient Boosting regression.
Calls \code{\link[xgboost:xgb.train]{xgboost::xgb.train()}} from package \CRANpkg{xgboost}.
}
\examples{
learner = mlr3::mlr_learners$get("regr.xgboost")
print(learner)

# available parameters:
as.data.table(learner$param_set)[, list(id, class, lower, upper, levels)]
}
\keyword{datasets}
