% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners-package}
\alias{mlr3learners}
\alias{mlr3learners-package}
\title{mlr3learners: Recommended Learners for 'mlr3'}
\description{
More learners are available in the \code{mlr3learners} repository on Github (\url{https://github.com/mlr3learners}).
There also is a wiki page listing all currently available custom learners (\url{https://github.com/mlr-org/mlr3learners/wiki/Extra-Learners}).
A guide on how to create custom learners is covered in the book: \url{https://mlr3book.mlr-org.com}.
Feel invited to contribute a missing learner to the \CRANpkg{mlr3} ecosystem!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3learners.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3learners}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3learners/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)

Authors:
\itemize{
  \item Quay Au \email{quayau@gmail.com} (0000-0002-5252-8902)
  \item Stefan Coors \email{mail@stefancoors.de} (0000-0002-7465-2146)
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (0000-0003-0748-6624)
}

}
