% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_glmnet.R
\docType{data}
\name{mlr_learners_classif.glmnet}
\alias{mlr_learners_classif.glmnet}
\alias{LearnerClassifGlmnet}
\title{GLM with Elastic Net Regularization Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifGlmnet
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for a generalized linear model with elastic net regularization implemented in \code{\link[glmnet:glmnet]{glmnet::glmnet()}} in package \CRANpkg{glmnet}.
}
\keyword{datasets}
