% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifLogReg.R
\name{mlr_learners_classif.log_reg}
\alias{mlr_learners_classif.log_reg}
\alias{LearnerClassifLogReg}
\title{Logistic Regression Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\description{
Classification via logistic regression.
Calls \code{\link[stats:glm]{stats::glm()}} with \code{family} set to \code{"binomial"}.
Argument \code{model} is set to \code{FALSE}.
}
\section{Construction}{
\preformatted{LearnerClassifLogReg$new()
mlr3::mlr_learners$get("classif.log_reg")
mlr3::lrn("classif.log_reg")
}
}

\section{Contrasts}{

To ensure reproducibility, this learner always uses the default contrasts:
\itemize{
\item \code{\link[=contr.treatment]{contr.treatment()}} for unordered factors, and
\item \code{\link[=contr.poly]{contr.poly()}} for ordered factors.
}

Setting the option \code{"contrasts"} does not have any effect.
Instead, set the respective hyperparameter or use \CRANpkg{mlr3pipelines} to create dummy features.
}

\examples{
learner = mlr3::lrn("classif.log_reg")
print(learner)

# available parameters:
learner$param_set$ids()
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
