% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_logreg.R
\docType{data}
\name{mlr_learners_classif.logreg}
\alias{mlr_learners_classif.logreg}
\alias{LearnerClassifLogReg}
\title{Classification Logistic Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifLogReg
}
\description{
A learner for a classification logistic regression implemented in \code{\link[stats:glm]{stats::glm()}}.
}
\keyword{datasets}
