% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_qda.R
\docType{data}
\name{mlr_learners_classif.qda}
\alias{mlr_learners_classif.qda}
\alias{LearnerClassifQDA}
\title{Quadratic Discriminant Analysis Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifQDA
}
\description{
Quadratic discriminant analysis.
Calls \code{\link[MASS:qda]{MASS::qda()}} from package \CRANpkg{MASS}.
}
\keyword{datasets}
