% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_svm.R
\docType{data}
\name{mlr_learners_classif.svm}
\alias{mlr_learners_classif.svm}
\alias{LearnerClassifSvm}
\title{Support Vector Machine}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifSvm
}
\description{
A learner for a classification support vector machine implemented in \code{\link[e1071:svm]{e1071::svm()}}.
}
\keyword{datasets}
