% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_classif.R
\docType{data}
\name{mlr_learners_classif.template}
\alias{mlr_learners_classif.template}
\alias{LearnerClassifTEMPLATE}
\title{\link{Your learner name} Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifTEMPLATE
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for \link{short description of your learner, mentioning the package it comes from}.
}
\keyword{datasets}
