% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_glmnet.R
\docType{data}
\name{mlr_learners_classif_glmnet}
\alias{mlr_learners_classif_glmnet}
\alias{LearnerClassifGlmnet}
\title{Classification GLM with Lasso or Elasticnet Regularization Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifGlmnet
}
\description{
A learner for a classification logistic regression implemented in \code{\link[glmnet:glmnet]{glmnet::glmnet()}}.
}
\keyword{datasets}
