% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_kknn.R
\docType{data}
\name{mlr_learners_classif_kknn}
\alias{mlr_learners_classif_kknn}
\alias{LearnerClassifKKNN}
\title{Classification k-Nearest-Neighbor Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifKKNN
}
\description{
A learner for a classification k-Nearest-Neighbor implemented in \code{\link[kknn:kknn]{kknn::kknn()}}.
}
\keyword{datasets}
