% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_ranger.R,
%   R/mlr_learners_regr_ranger.R
\docType{data}
\name{mlr_learners_classif_ranger}
\alias{mlr_learners_classif_ranger}
\alias{LearnerClassifRanger}
\alias{LearnerRegrRanger}
\title{Classification Ranger Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifRanger

LearnerRegrRanger
}
\description{
A learner for a classification random forest implemented in \code{\link[ranger:ranger]{ranger::ranger()}}.

A learner for a regression random forest implemented in \code{\link[ranger:ranger]{ranger::ranger()}}.
}
\keyword{datasets}
