% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_xgboost.R
\docType{data}
\name{mlr_learners_classif_xgboost}
\alias{mlr_learners_classif_xgboost}
\alias{LearnerClassifXgboost}
\title{Classification xgboost Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifXgboost
}
\description{
An eXtreme Gradient Boosting learner for classification, implemented in \code{\link[xgboost:xgb.train]{xgboost::xgb.train()}}.
}
\keyword{datasets}
