% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrKKNN.R
\name{mlr_learners_regr.kknn}
\alias{mlr_learners_regr.kknn}
\alias{LearnerRegrKKNN}
\title{k-Nearest-Neighbor Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\description{
k-Nearest-Neighbor regression.
Calls \code{\link[kknn:kknn]{kknn::kknn()}} from package \CRANpkg{kknn}.
}
\section{Construction}{
\preformatted{LearnerRegrKKNN$new()
mlr3::mlr_learners$get("regr.kknn")
mlr3::lrn("regr.kknn")
}
}

\examples{
learner = mlr3::lrn("regr.kknn")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
\cite{mlr3learners}{hechenbichler_2004}

\cite{mlr3learners}{samworth_2012}

\cite{mlr3learners}{cover_1967}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
