% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_regr_kriging.R
\docType{data}
\name{mlr_learners_regr.kriging}
\alias{mlr_learners_regr.kriging}
\alias{LearnerRegrKriging}
\title{Kriging Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrKriging
}
\description{
Kriging regression model.
Calls \code{\link[DiceKriging:km]{DiceKriging::km()}} from package \CRANpkg{DiceKriging}.
}
\keyword{datasets}
