% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_regr_svm.R
\docType{data}
\name{mlr_learners_regr.svm}
\alias{mlr_learners_regr.svm}
\alias{LearnerRegrSvm}
\title{Support Vector Machine}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrSvm
}
\description{
A learner for a regression support vector machine implemented in \code{\link[e1071:svm]{e1071::svm()}}.
}
\keyword{datasets}
