% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_regr.R
\docType{data}
\name{mlr_learners_regr.template}
\alias{mlr_learners_regr.template}
\alias{LearnerRegrTEMPLATE}
\title{\link{Your learner name} Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrTEMPLATE
}
\description{
A \link[mlr3:LearnerRegr]{mlr3::LearnerRegr} for \link{short description of your learner, mentioning the package it comes from}.
}
\keyword{datasets}
