% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_regr_kknn.R
\docType{data}
\name{mlr_learners_regr_kknn}
\alias{mlr_learners_regr_kknn}
\alias{LearnerRegrKKNN}
\title{Regression k-Nearest-Neighbor Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrKKNN
}
\description{
A learner for a regression k-Nearest-Neighbor implemented in \code{\link[kknn:kknn]{kknn::kknn()}}.
}
\keyword{datasets}
