% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifGlmnet.R
\docType{data}
\name{LearnerClassifGlmnet}
\alias{LearnerClassifGlmnet}
\alias{mlr_learners_classif.glmnet}
\title{GLM with Elastic Net Regularization Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifGlmnet
}
\description{
Generalized linear models with elastic net regularization.
Calls \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} from package \CRANpkg{glmnet}.
}
\examples{
learner = mlr3::mlr_learners$get("classif.glmnet")
print(learner)

# available parameters:
tab = data.table::as.data.table(learner$param_set)
tab[, list(id, class, lower, upper, levels)]
}
\keyword{datasets}
