% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifKKNN.R
\docType{data}
\name{LearnerClassifKKNN}
\alias{LearnerClassifKKNN}
\alias{mlr_learners_classif.kknn}
\title{k-Nearest-Neighbor Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\description{
k-Nearest-Neighbor classification.
Calls \code{\link[kknn:kknn]{kknn::kknn()}} from package \CRANpkg{kknn}.
}
\section{Construction}{
\preformatted{LearnerClassifKKNN$new()
mlr3::mlr_learners$get("classif.kknn")
mlr3::lrn("classif.kknn")
}
}

\examples{
learner = mlr3::lrn("classif.kknn")
print(learner)

# available parameters:
learner$param_set$ids()
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
