% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifLDA.R
\docType{data}
\name{LearnerClassifLDA}
\alias{LearnerClassifLDA}
\alias{mlr_learners_classif.lda}
\title{Linear Discriminant Analysis Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifLDA
}
\description{
Linear discriminant analysis.
Calls \code{\link[MASS:lda]{MASS::lda()}} from package \CRANpkg{MASS}.

Venables, W. N. & Ripley, B. D. (2002).
Modern Applied Statistics with S.
Fourth Edition. Springer, New York. ISBN 0-387-95457-0
}
\examples{
learner = mlr3::mlr_learners$get("classif.lda")
print(learner)

# available parameters:
learner$param_set$ids()
}
\keyword{datasets}
