% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifLDA.R
\docType{data}
\name{LearnerClassifLDA}
\alias{LearnerClassifLDA}
\alias{mlr_learners_classif.lda}
\title{Linear Discriminant Analysis Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifLDA
}
\description{
Linear discriminant analysis.
Calls \code{\link[MASS:lda]{MASS::lda()}} from package \CRANpkg{MASS}.
}
\examples{
learner = mlr3::mlr_learners$get("classif.lda")
print(learner)

# available parameters:
tab = data.table::as.data.table(learner$param_set)
tab[, list(id, class, lower, upper, levels)]
}
\keyword{datasets}
