% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifLogReg.R
\docType{data}
\name{LearnerClassifLogReg}
\alias{LearnerClassifLogReg}
\alias{mlr_learners_classif.log_reg}
\title{Logistic Regression Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifLogReg
}
\description{
Classification via logistic regression.
Calls \code{\link[stats:glm]{stats::glm()}}.
}
\keyword{datasets}
