% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifQDA.R
\docType{data}
\name{LearnerClassifQDA}
\alias{LearnerClassifQDA}
\alias{mlr_learners_classif.qda}
\title{Quadratic Discriminant Analysis Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifQDA
}
\description{
Quadratic discriminant analysis.
Calls \code{\link[MASS:qda]{MASS::qda()}} from package \CRANpkg{MASS}.
}
\examples{
learner = mlr3::mlr_learners$get("classif.qda")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Venables, W. N. & Ripley, B. D (2002).
Modern Applied Statistics with S.
Fourth Edition. Springer, New York. ISBN 0-387-95457-0
}
\keyword{datasets}
