% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifRanger.R
\docType{data}
\name{LearnerClassifRanger}
\alias{LearnerClassifRanger}
\alias{mlr_learners_classif.ranger}
\title{Ranger Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifRanger
}
\description{
Random classification forest.
Calls \code{\link[ranger:ranger]{ranger::ranger()}} from package \CRANpkg{ranger}.
}
\examples{
learner = mlr3::mlr_learners$get("classif.ranger")
print(learner)
print(learner$param_set)
}
\keyword{datasets}
