% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifRanger.R
\docType{data}
\name{LearnerClassifRanger}
\alias{LearnerClassifRanger}
\alias{mlr_learners_classif.ranger}
\title{Ranger Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\description{
Random classification forest.
Calls \code{\link[ranger:ranger]{ranger::ranger()}} from package \CRANpkg{ranger}.

\insertNoCite{breiman_2001}{mlr3learners}
\insertNoCite{wright_2017}{mlr3learners}
}
\section{Construction}{
\preformatted{LearnerClassifRanger$new()
mlr3::mlr_learners$get("classif.ranger")
mlr3::lrn("classif.ranger")
}
}

\examples{
learner = mlr3::lrn("classif.ranger")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
\insertAllCited{}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
