% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifSVM.R
\docType{data}
\name{LearnerClassifSVM}
\alias{LearnerClassifSVM}
\alias{mlr_learners_classif.svm}
\title{Support Vector Machine}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\description{
A learner for a classification support vector machine implemented in \code{\link[e1071:svm]{e1071::svm()}}.
}
\section{Construction}{
\preformatted{LearnerClassifSVM$new()
mlr3::mlr_learners$get("classif.svm")
mlr3::lrn("classif.svm")
}
}

\examples{
learner = mlr3::lrn("classif.svm")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
\cite{mlr3learners}{cortes_1995}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
