% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifXgboost.R
\docType{data}
\name{LearnerClassifXgboost}
\alias{LearnerClassifXgboost}
\alias{mlr_learners_classif.xgboost}
\title{Extreme Gradiant Boosting Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifXgboost
}
\description{
eXtreme Gradient Boosting classification.
Calls \code{\link[xgboost:xgb.train]{xgboost::xgb.train()}} from package \CRANpkg{xgboost}.
}
\examples{
learner = mlr3::mlr_learners$get("classif.xgboost")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Tianqi Chen and Carlos Guestrin (2016).
XGBoost: A Scalable Tree Boosting System.
22nd SIGKDD Conference on Knowledge Discovery and Data Mining.
\doi{10.1145/2939672.2939785}.
}
\keyword{datasets}
