% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrGlmnet.R
\docType{data}
\name{LearnerRegrGlmnet}
\alias{LearnerRegrGlmnet}
\alias{mlr_learners_regr.glmnet}
\title{GLM with Elastic Net Regularization Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\description{
Generalized linear models with elastic net regularization.
Calls \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} from package \CRANpkg{glmnet}.
Hyperparameter \code{family} is set to \code{"gaussian"}.

\insertNoCite{friedman_2010}{mlr3learners}
}
\section{Construction}{
\preformatted{LearnerRegrGlmnet$new()
mlr3::mlr_learners$get("regr.glmnet")
mlr3::lrn("regr.glmnet")
}
}

\examples{
learner = mlr3::lrn("regr.glmnet")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
\insertAllCited{}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
