% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrKKNN.R
\docType{data}
\name{LearnerRegrKKNN}
\alias{LearnerRegrKKNN}
\alias{mlr_learners_regr.kknn}
\title{k-Nearest-Neighbor Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\description{
k-Nearest-Neighbor regression.
Calls \code{\link[kknn:kknn]{kknn::kknn()}} from package \CRANpkg{kknn}.

\insertNoCite{cover_1967}{mlr3learners}
\insertNoCite{samworth_2012}{mlr3learners}
\insertNoCite{hechenbichler_2004}{mlr3learners}
}
\section{Construction}{
\preformatted{LearnerRegrKKNN$new()
mlr3::mlr_learners$get("regr.kknn")
mlr3::lrn("regr.kknn")
}
}

\examples{
learner = mlr3::lrn("regr.kknn")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
\insertAllCited{}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
