% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrKKNN.R
\docType{data}
\name{LearnerRegrKKNN}
\alias{LearnerRegrKKNN}
\alias{mlr_learners_regr.kknn}
\title{k-Nearest-Neighbor Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrKKNN
}
\description{
k-Nearest-Neighbor regression.
Calls \code{\link[kknn:kknn]{kknn::kknn()}} from package \CRANpkg{kknn}.
}
\examples{
learner = mlr3::mlr_learners$get("regr.kknn")
print(learner)

# available parameters:
learner$param_set$ids()
}
\keyword{datasets}
