% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrKM.R
\docType{data}
\name{LearnerRegrKM}
\alias{LearnerRegrKM}
\alias{mlr_learners_regr.km}
\title{Kriging Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrKM
}
\description{
Kriging regression.
Calls \code{\link[DiceKriging:km]{DiceKriging::km()}} from package \CRANpkg{DiceKriging}.
\itemize{
\item The predict type hyperparameter "type" defaults to "SK" (simple Kriging).
\item The additional hyperparameter \code{nugget.stability} is used to overwrite the hyperparameter \code{nugget} with \code{nugget.stability * var(y)} before training to improve the numerical stability.
We recommend a value of \code{1e-8}.
\item The additional hyperparameter \code{jitter} can be set to add \code{N(0, [jitter])}-distributed noise to the data before prediction to avoid perfect interpolation. We recommend a value if \code{1e-12}.
}
}
\examples{
learner = mlr3::mlr_learners$get("regr.km")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Olivier Roustant, David Ginsbourger, Yves Deville (2012).
DiceKriging, DiceOptim: Two R Packages for the Analysis of Computer Experiments by Kriging-Based Metamodeling and Optimization.
Journal of Statistical Software, 51(1), 1-55.
\doi{10.18637/jss.v051.i01}.
}
\keyword{datasets}
