% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrLM.R
\docType{data}
\name{LearnerRegrLM}
\alias{LearnerRegrLM}
\alias{mlr_learners_regr.lm}
\title{Linear Model Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrLM
}
\description{
Ordinary linear regression.
Calls \code{\link[stats:lm]{stats::lm()}}.
}
\examples{
learner = mlr3::mlr_learners$get("regr.lm")
print(learner)

# available parameters:
learner$param_set$ids()
}
\keyword{datasets}
