% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrXgboost.R
\docType{data}
\name{LearnerRegrXgboost}
\alias{LearnerRegrXgboost}
\alias{mlr_learners_regr.xgboost}
\title{Extreme Gradient Boosting Regressiong Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrXgboost
}
\description{
eXtreme Gradient Boosting regression.
Calls \code{\link[xgboost:xgb.train]{xgboost::xgb.train()}} from package \CRANpkg{xgboost}.
}
\examples{
learner = mlr3::mlr_learners$get("regr.xgboost")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Tianqi Chen and Carlos Guestrin (2016).
XGBoost: A Scalable Tree Boosting System
22nd SIGKDD Conference on Knowledge Discovery and Data Mining,
https://arxiv.org/abs/1603.02754
}
\keyword{datasets}
