% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrXgboost.R
\docType{data}
\name{LearnerRegrXgboost}
\alias{LearnerRegrXgboost}
\alias{mlr_learners_regr.xgboost}
\title{Extreme Gradient Boosting Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\description{
eXtreme Gradient Boosting regression.
Calls \code{\link[xgboost:xgb.train]{xgboost::xgb.train()}} from package \CRANpkg{xgboost}.
}
\section{Construction}{
\preformatted{LearnerRegrXgboost$new()
mlr3::mlr_learners$get("regr.xgboost")
mlr3::lrn("regr.xgboost")
}
}

\examples{
learner = mlr3::lrn("regr.xgboost")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
\cite{mlr3learners}{chen_2016}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
