% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners-package}
\alias{mlr3learners}
\alias{mlr3learners-package}
\title{mlr3learners: Recommended Learners for 'mlr3'}
\description{
More learners are implemented in the \href{https://github.com/mlr-org/mlr3extralearners}{mlr3extralearners package}.
A guide on how to create custom learners is covered in the book:
\url{https://mlr3book.mlr-org.com}.
Feel invited to contribute a missing learner to the \CRANpkg{mlr3} ecosystem!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3learners.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3learners}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3learners/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
  \item Quay Au \email{quayau@gmail.com} (\href{https://orcid.org/0000-0002-5252-8902}{ORCID})
  \item Stefan Coors \email{mail@stefancoors.de} (\href{https://orcid.org/0000-0002-7465-2146}{ORCID})
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})
  \item John Zobolas \email{bblodfon@gmail.com} (\href{https://orcid.org/0000-0002-3609-8674}{ORCID})
}

}
