% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners-package}
\alias{mlr3learners}
\alias{mlr3learners-package}
\title{mlr3learners: Recommended Learners for 'mlr3'}
\description{
Recommended Learners for 'mlr3'. Extends 'mlr3'
    with interfaces to the most popular machine learning packages on CRAN.
}
\section{Classification Learners}{

\tabular{lll}{
\link{classif.glmnet}      \tab \CRANpkg{glmnet}  \tab Penalized Logistic Regression
\link{classif.kknn}        \tab \CRANpkg{kknn}    \tab k-Nearest Neighbors
\link{classif.lda}         \tab \CRANpkg{MASS}    \tab Linear Discriminant Analysis
\link{classif.log_reg}     \tab \CRANpkg{stats}   \tab Logistic Regression
\link{classif.naive_bayes} \tab \CRANpkg{e1071}   \tab Naive Bayes
\link{classif.qda}         \tab \CRANpkg{MASS}    \tab Quadratic Discriminant Analysis
\link{classif.ranger}      \tab \CRANpkg{ranger}  \tab Random Classification Forest
\link{classif.svm}         \tab \CRANpkg{e1071}   \tab Support Vector Machine
\link{classif.xgboost}     \tab \CRANpkg{xgboost} \tab Gradient Boosting
}
}

\section{Regression Learners}{

\tabular{lll}{
\link{regr.glmnet}  \tab \CRANpkg{glmnet}      \tab Penalized Linear Regression
\link{regr.kknn}    \tab \CRANpkg{kknn}        \tab k-Nearest Neighbors
\link{regr.km}      \tab \CRANpkg{DiceKriging} \tab Kriging
\link{regr.lm}      \tab \CRANpkg{stats}       \tab Linear Regression
\link{regr.ranger}  \tab \CRANpkg{ranger}      \tab Random Regression Forest
\link{regr.svm}     \tab \CRANpkg{e1071}       \tab Support Vector Machine
\link{regr.xgboost} \tab \CRANpkg{xgboost}     \tab Gradient Boosting
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3learners.mlr-org.com}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3learners/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)

Authors:
\itemize{
  \item Quay Au \email{quayau@gmail.com} (0000-0002-5252-8902)
  \item Stefan Coors \email{mail@stefancoors.de} (0000-0002-7465-2146)
}

}
