% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners-package}
\alias{mlr3learners}
\alias{mlr3learners-package}
\title{mlr3learners: Recommended Learners for 'mlr3'}
\description{
Recommended Learners for 'mlr3'. Extends 'mlr3'
    with interfaces to essential machine learning packages on CRAN.
    This includes, but is not limited to: (penalized) linear and logistic
    regression, linear and quadratic discriminant analysis, k-nearest neighbors,
    naive Bayes, support vector machines, and gradient boosting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3learners.mlr-org.com}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3learners/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)

Authors:
\itemize{
  \item Quay Au \email{quayau@gmail.com} (0000-0002-5252-8902)
  \item Stefan Coors \email{mail@stefancoors.de} (0000-0002-7465-2146)
}

}
