% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_naive_bayes.R
\docType{data}
\name{mlr_learners_classif.naive_bayes}
\alias{mlr_learners_classif.naive_bayes}
\alias{LearnerClassifNaiveBayes}
\title{Naive Bayes Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifNaiveBayes
}
\description{
Naive Bayes classification.
Calls \code{\link[e1071:naiveBayes]{e1071::naiveBayes()}} from package \CRANpkg{e1071}.
}
\keyword{datasets}
