% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_ranger.R
\docType{data}
\name{mlr_learners_classif.ranger}
\alias{mlr_learners_classif.ranger}
\alias{LearnerClassifRanger}
\title{Ranger Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifRanger
}
\description{
Random classification forest.
Calls \code{\link[ranger:ranger]{ranger::ranger()}} from package \CRANpkg{ranger}.
}
\keyword{datasets}
