% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_xgboost.R
\docType{data}
\name{mlr_learners_classif.xgboost}
\alias{mlr_learners_classif.xgboost}
\alias{LearnerClassifXgboost}
\title{Extreme Gradiant Boosting Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifXgboost
}
\description{
An \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for eXtreme Gradient Boosting learner implemented in \code{\link[xgboost:xgb.train]{xgboost::xgb.train()}} in package \CRANpkg{xgboost}.
}
\keyword{datasets}
