% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_ranger.R
\docType{data}
\name{mlr_learners_classif_ranger}
\alias{mlr_learners_classif_ranger}
\alias{LearnerClassifRanger}
\title{Classification Ranger Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifRanger
}
\description{
A learner for a classification random forest implemented in \code{\link[ranger:ranger]{ranger::ranger()}}.
}
\keyword{datasets}
