% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_regr_glmnet.R
\docType{data}
\name{mlr_learners_regr.glmnet}
\alias{mlr_learners_regr.glmnet}
\alias{LearnerRegrGlmnet}
\title{GLM with Elastic Net Regularization Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrGlmnet
}
\description{
A \link[mlr3:LearnerRegr]{mlr3::LearnerRegr} for a regression GLM with elastic net regularization implemented in \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} in package \CRANpkg{glmnet}.
}
\keyword{datasets}
