% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_regr_glmnet.R
\docType{data}
\name{mlr_learners_regr.glmnet}
\alias{mlr_learners_regr.glmnet}
\alias{LearnerRegrGlmnet}
\title{Regression GLM with Elastic net Regularization}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrGlmnet
}
\description{
A learner for a regression GLM with elastic net regularization implemented in \code{\link[glmnet:glmnet]{glmnet::glmnet()}}.
}
\keyword{datasets}
