% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrGlmnet.R
\name{mlr_learners_regr.glmnet}
\alias{mlr_learners_regr.glmnet}
\alias{LearnerRegrGlmnet}
\title{GLM with Elastic Net Regularization Regression Learner}
\description{
Generalized linear models with elastic net regularization.
Calls \code{\link[glmnet:glmnet]{glmnet::glmnet()}} from package \CRANpkg{glmnet}.

The default for hyperparameter \code{family} is set to \code{"gaussian"}.
}
\details{
Caution: This learner is different to learners calling \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
in that it does not use the internal optimization of parameter \code{lambda}.
Instead, \code{lambda} needs to be tuned by the user (e.g., via \CRANpkg{mlr3tuning}).
When \code{lambda} is tuned, the \code{glmnet} will be trained for each tuning iteration.
While fitting the whole path of \code{lambda}s would be more efficient, as is done
by default in \code{\link[glmnet:glmnet]{glmnet::glmnet()}}, tuning/selecting the parameter at prediction time
(using parameter \code{s}) is currently not supported in \CRANpkg{mlr3}
(at least not in efficient manner).
Tuning the \code{s} parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective
auto-tuning counterparts in \code{\link[=mlr_learners_classif.cv_glmnet]{mlr_learners_classif.cv_glmnet()}} or
\code{\link[=mlr_learners_regr.cv_glmnet]{mlr_learners_regr.cv_glmnet()}}.
However, in some situations this is not applicable, usually when data are
imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires
custom resampling strategies (blocked design, stratification).
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("regr.glmnet")
lrn("regr.glmnet")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

, * Task type: \dQuote{regr}, * Predict Types: \dQuote{response}, * Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, * Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3learners}, \CRANpkg{glmnet}
}

\section{Parameters}{

, |Id                   |Type      |Default  |Levels                  |Range                                |, |:--------------------|:---------|:--------|:-----------------------|:------------------------------------|, |alignment            |character |lambda   |lambda, fraction        |-                                    |, |alpha                |numeric   |1        |                        |\eqn{[0, 1]}{[0, 1]}                 |, |big                  |numeric   |9.9e+35  |                        |\eqn{(-\infty, \infty)}{(-Inf, Inf)} |, |devmax               |numeric   |0.999    |                        |\eqn{[0, 1]}{[0, 1]}                 |, |dfmax                |integer   |-        |                        |\eqn{[0, \infty)}{[0, Inf)}          |, |eps                  |numeric   |1e-06    |                        |\eqn{[0, 1]}{[0, 1]}                 |, |epsnr                |numeric   |1e-08    |                        |\eqn{[0, 1]}{[0, 1]}                 |, |exact                |logical   |FALSE    |TRUE, FALSE             |-                                    |, |exclude              |integer   |-        |                        |\eqn{[1, \infty)}{[1, Inf)}          |, |exmx                 |numeric   |250      |                        |\eqn{(-\infty, \infty)}{(-Inf, Inf)} |, |family               |character |gaussian |gaussian, poisson       |-                                    |, |fdev                 |numeric   |1e-05    |                        |\eqn{[0, 1]}{[0, 1]}                 |, |gamma                |numeric   |1        |                        |\eqn{(-\infty, \infty)}{(-Inf, Inf)} |, |grouped              |logical   |TRUE     |TRUE, FALSE             |-                                    |, |intercept            |logical   |TRUE     |TRUE, FALSE             |-                                    |, |keep                 |logical   |FALSE    |TRUE, FALSE             |-                                    |, |lambda               |untyped   |-        |                        |-                                    |, |lambda.min.ratio     |numeric   |-        |                        |\eqn{[0, 1]}{[0, 1]}                 |, |lower.limits         |untyped   |-        |                        |-                                    |, |maxit                |integer   |100000   |                        |\eqn{[1, \infty)}{[1, Inf)}          |, |mnlam                |integer   |5        |                        |\eqn{[1, \infty)}{[1, Inf)}          |, |mxit                 |integer   |100      |                        |\eqn{[1, \infty)}{[1, Inf)}          |, |mxitnr               |integer   |25       |                        |\eqn{[1, \infty)}{[1, Inf)}          |, |newoffset            |untyped   |-        |                        |-                                    |, |nlambda              |integer   |100      |                        |\eqn{[1, \infty)}{[1, Inf)}          |, |offset               |untyped   |         |                        |-                                    |, |parallel             |logical   |FALSE    |TRUE, FALSE             |-                                    |, |penalty.factor       |untyped   |-        |                        |-                                    |, |pmax                 |integer   |-        |                        |\eqn{[0, \infty)}{[0, Inf)}          |, |pmin                 |numeric   |1e-09    |                        |\eqn{[0, 1]}{[0, 1]}                 |, |prec                 |numeric   |1e-10    |                        |\eqn{(-\infty, \infty)}{(-Inf, Inf)} |, |relax                |logical   |FALSE    |TRUE, FALSE             |-                                    |, |s                    |numeric   |0.01     |                        |\eqn{[0, \infty)}{[0, Inf)}          |, |standardize          |logical   |TRUE     |TRUE, FALSE             |-                                    |, |standardize.response |logical   |FALSE    |TRUE, FALSE             |-                                    |, |thresh               |numeric   |1e-07    |                        |\eqn{[0, \infty)}{[0, Inf)}          |, |trace.it             |integer   |0        |                        |\eqn{[0, 1]}{[0, 1]}                 |, |type.gaussian        |character |-        |covariance, naive       |-                                    |, |type.logistic        |character |-        |Newton, modified.Newton |-                                    |, |type.multinomial     |character |-        |ungrouped, grouped      |-                                    |, |upper.limits         |untyped   |-        |                        |-                                    |
}

\examples{
if (requireNamespace("glmnet", quietly = TRUE)) {
  learner = mlr3::lrn("regr.glmnet")
  print(learner)

  # available parameters:
learner$param_set$ids()
}
}
\references{
Friedman J, Hastie T, Tibshirani R (2010).
\dQuote{Regularization Paths for Generalized Linear Models via Coordinate Descent.}
\emph{Journal of Statistical Software}, \bold{33}(1), 1--22.
\doi{10.18637/jss.v033.i01}.
}
\seealso{
\itemize{
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}:
\url{https://mlr3book.mlr-org.com/basics.html#learners}
\item Package \href{https://github.com/mlr-org/mlr3extralearners}{mlr3extralearners} for more learners.
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[=Learner]{Learners}: \link{mlr_learners}
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item Extension packages for additional task types:
\itemize{
\item \CRANpkg{mlr3proba} for probabilistic supervised regression and survival analysis.
\item \CRANpkg{mlr3cluster} for unsupervised clustering.
}
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces}
for established default tuning spaces.
}

Other Learner: 
\code{\link{mlr_learners_classif.cv_glmnet}},
\code{\link{mlr_learners_classif.glmnet}},
\code{\link{mlr_learners_classif.kknn}},
\code{\link{mlr_learners_classif.lda}},
\code{\link{mlr_learners_classif.log_reg}},
\code{\link{mlr_learners_classif.multinom}},
\code{\link{mlr_learners_classif.naive_bayes}},
\code{\link{mlr_learners_classif.nnet}},
\code{\link{mlr_learners_classif.qda}},
\code{\link{mlr_learners_classif.ranger}},
\code{\link{mlr_learners_classif.svm}},
\code{\link{mlr_learners_classif.xgboost}},
\code{\link{mlr_learners_regr.cv_glmnet}},
\code{\link{mlr_learners_regr.kknn}},
\code{\link{mlr_learners_regr.km}},
\code{\link{mlr_learners_regr.lm}},
\code{\link{mlr_learners_regr.ranger}},
\code{\link{mlr_learners_regr.svm}},
\code{\link{mlr_learners_regr.xgboost}}
}
\concept{Learner}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrGlmnet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerRegrGlmnet-new}{\code{LearnerRegrGlmnet$new()}}
\item \href{#method-LearnerRegrGlmnet-selected_features}{\code{LearnerRegrGlmnet$selected_features()}}
\item \href{#method-LearnerRegrGlmnet-clone}{\code{LearnerRegrGlmnet$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrGlmnet-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrGlmnet-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrGlmnet$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrGlmnet-selected_features"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrGlmnet-selected_features}{}}}
\subsection{Method \code{selected_features()}}{
Returns the set of selected features as reported by \code{\link[glmnet:predict.glmnet]{glmnet::predict.glmnet()}}
with \code{type} set to \code{"nonzero"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrGlmnet$selected_features(lambda = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{(\code{numeric(1)})\cr
Custom \code{lambda}, defaults to the active lambda depending on parameter set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{character()}) of feature names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrGlmnet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrGlmnet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrGlmnet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
