% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrGlmnet.R
\name{mlr_learners_regr.glmnet}
\alias{mlr_learners_regr.glmnet}
\alias{LearnerRegrGlmnet}
\title{GLM with Elastic Net Regularization Regression Learner}
\description{
Generalized linear models with elastic net regularization.
Calls \code{\link[glmnet:glmnet]{glmnet::glmnet()}} from package \CRANpkg{glmnet}.

The default for hyperparameter \code{family} is set to \code{"gaussian"}.
}
\details{
Caution: This learner is different to learners calling \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
in that it does not use the internal optimization of parameter \code{lambda}.
Instead, \code{lambda} needs to be tuned by the user (e.g., via \CRANpkg{mlr3tuning}).
When \code{lambda} is tuned, the \code{glmnet} will be trained for each tuning iteration.
While fitting the whole path of \code{lambda}s would be more efficient, as is done
by default in \code{\link[glmnet:glmnet]{glmnet::glmnet()}}, tuning/selecting the parameter at prediction time
(using parameter \code{s}) is currently not supported in \CRANpkg{mlr3}
(at least not in efficient manner).
Tuning the \code{s} parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective
auto-tuning counterparts in \code{\link[=mlr_learners_classif.cv_glmnet]{mlr_learners_classif.cv_glmnet()}},
\code{\link[=mlr_learners_regr.cv_glmnet]{mlr_learners_regr.cv_glmnet()}}, or \code{\link[=mlr_learners_surv.cv_glmnet]{mlr_learners_surv.cv_glmnet()}}.
However, in some situations this is not applicable, usually when data are
imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires
custom resampling strategies (blocked design, stratification).
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:\preformatted{mlr_learners$get("regr.glmnet")
lrn("regr.glmnet")
}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Predict Types: \dQuote{response}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3learners}, \CRANpkg{glmnet}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Range \tab Levels \cr
   alignment \tab character \tab lambda \tab - \tab lambda, fraction \cr
   alpha \tab numeric \tab 1 \tab \eqn{[0, 1]}{[0, 1]} \tab - \cr
   big \tab numeric \tab 9.9e+35 \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \tab - \cr
   devmax \tab numeric \tab 0.999 \tab \eqn{[0, 1]}{[0, 1]} \tab - \cr
   dfmax \tab integer \tab - \tab \eqn{[0, \infty)}{[0, Inf)} \tab - \cr
   eps \tab numeric \tab 1e-06 \tab \eqn{[0, 1]}{[0, 1]} \tab - \cr
   epsnr \tab numeric \tab 1e-08 \tab \eqn{[0, 1]}{[0, 1]} \tab - \cr
   exact \tab logical \tab FALSE \tab - \tab TRUE, FALSE \cr
   exclude \tab integer \tab - \tab \eqn{[1, \infty)}{[1, Inf)} \tab - \cr
   exmx \tab numeric \tab 250 \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \tab - \cr
   family \tab character \tab gaussian \tab - \tab gaussian, poisson \cr
   fdev \tab numeric \tab 1e-05 \tab \eqn{[0, 1]}{[0, 1]} \tab - \cr
   gamma \tab numeric \tab 1 \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \tab - \cr
   grouped \tab logical \tab TRUE \tab - \tab TRUE, FALSE \cr
   intercept \tab logical \tab TRUE \tab - \tab TRUE, FALSE \cr
   keep \tab logical \tab FALSE \tab - \tab TRUE, FALSE \cr
   lambda \tab list \tab - \tab - \tab - \cr
   lambda.min.ratio \tab numeric \tab - \tab \eqn{[0, 1]}{[0, 1]} \tab - \cr
   lower.limits \tab list \tab - \tab - \tab - \cr
   maxit \tab integer \tab 100000 \tab \eqn{[1, \infty)}{[1, Inf)} \tab - \cr
   mnlam \tab integer \tab 5 \tab \eqn{[1, \infty)}{[1, Inf)} \tab - \cr
   mxit \tab integer \tab 100 \tab \eqn{[1, \infty)}{[1, Inf)} \tab - \cr
   mxitnr \tab integer \tab 25 \tab \eqn{[1, \infty)}{[1, Inf)} \tab - \cr
   newoffset \tab list \tab - \tab - \tab - \cr
   nlambda \tab integer \tab 100 \tab \eqn{[1, \infty)}{[1, Inf)} \tab - \cr
   offset \tab list \tab NULL \tab - \tab - \cr
   parallel \tab logical \tab FALSE \tab - \tab TRUE, FALSE \cr
   penalty.factor \tab list \tab - \tab - \tab - \cr
   pmax \tab integer \tab - \tab \eqn{[0, \infty)}{[0, Inf)} \tab - \cr
   pmin \tab numeric \tab 1e-09 \tab \eqn{[0, 1]}{[0, 1]} \tab - \cr
   prec \tab numeric \tab 1e-10 \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \tab - \cr
   relax \tab logical \tab FALSE \tab - \tab TRUE, FALSE \cr
   s \tab numeric \tab 0.01 \tab \eqn{[0, \infty)}{[0, Inf)} \tab - \cr
   standardize \tab logical \tab TRUE \tab - \tab TRUE, FALSE \cr
   standardize.response \tab logical \tab FALSE \tab - \tab TRUE, FALSE \cr
   thresh \tab numeric \tab 1e-07 \tab \eqn{[0, \infty)}{[0, Inf)} \tab - \cr
   trace.it \tab integer \tab 0 \tab \eqn{[0, 1]}{[0, 1]} \tab - \cr
   type.gaussian \tab character \tab - \tab - \tab covariance, naive \cr
   type.logistic \tab character \tab - \tab - \tab Newton, modified.Newton \cr
   type.multinomial \tab character \tab - \tab - \tab ungrouped, grouped \cr
   upper.limits \tab list \tab - \tab - \tab - \cr
}
}

\examples{
if (requireNamespace("glmnet", quietly = TRUE)) {
  learner = mlr3::lrn("regr.glmnet")
  print(learner)

  # available parameters:
learner$param_set$ids()
}
}
\references{
Friedman J, Hastie T, Tibshirani R (2010).
\dQuote{Regularization Paths for Generalized Linear Models via Coordinate Descent.}
\emph{Journal of Statistical Software}, \bold{33}(1), 1--22.
\doi{10.18637/jss.v033.i01}.
}
\seealso{
\itemize{
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}:
\url{https://mlr3book.mlr-org.com/basics.html#learners}
\item Package \href{https://github.com/mlr-org/mlr3extralearners}{mlr3extralearners} for more learners.
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[=Learner]{Learners}: \link{mlr_learners}
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item Extension packages for additional task types:
\itemize{
\item \CRANpkg{mlr3proba} for probabilistic supervised regression and survival analysis.
\item \CRANpkg{mlr3cluster} for unsupervised clustering.
}
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces}
for established default tuning spaces.
}

Other Learner: 
\code{\link{mlr_learners_classif.cv_glmnet}},
\code{\link{mlr_learners_classif.glmnet}},
\code{\link{mlr_learners_classif.kknn}},
\code{\link{mlr_learners_classif.lda}},
\code{\link{mlr_learners_classif.log_reg}},
\code{\link{mlr_learners_classif.multinom}},
\code{\link{mlr_learners_classif.naive_bayes}},
\code{\link{mlr_learners_classif.nnet}},
\code{\link{mlr_learners_classif.qda}},
\code{\link{mlr_learners_classif.ranger}},
\code{\link{mlr_learners_classif.svm}},
\code{\link{mlr_learners_classif.xgboost}},
\code{\link{mlr_learners_regr.cv_glmnet}},
\code{\link{mlr_learners_regr.kknn}},
\code{\link{mlr_learners_regr.km}},
\code{\link{mlr_learners_regr.lm}},
\code{\link{mlr_learners_regr.ranger}},
\code{\link{mlr_learners_regr.svm}},
\code{\link{mlr_learners_regr.xgboost}},
\code{\link{mlr_learners_surv.cv_glmnet}},
\code{\link{mlr_learners_surv.glmnet}},
\code{\link{mlr_learners_surv.ranger}},
\code{\link{mlr_learners_surv.xgboost}}
}
\concept{Learner}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrGlmnet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerRegrGlmnet$new()}}
\item \href{#method-selected_features}{\code{LearnerRegrGlmnet$selected_features()}}
\item \href{#method-clone}{\code{LearnerRegrGlmnet$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner">}\href{../../mlr3/html/Learner.html#method-base_learner}{\code{mlr3::Learner$base_learner()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrGlmnet$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-selected_features"></a>}}
\if{latex}{\out{\hypertarget{method-selected_features}{}}}
\subsection{Method \code{selected_features()}}{
Returns the set of selected features as reported by \code{\link[glmnet:predict.glmnet]{glmnet::predict.glmnet()}}
with \code{type} set to \code{"nonzero"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrGlmnet$selected_features(lambda = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{(\code{numeric(1)})\cr
Custom \code{lambda}, defaults to the active lambda depending on parameter set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{character()}) of feature names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrGlmnet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
