% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_regr_kknn.R
\docType{data}
\name{mlr_learners_regr.kknn}
\alias{mlr_learners_regr.kknn}
\alias{LearnerRegrKKNN}
\title{k-Nearest-Neighbor Regression Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrKKNN
}
\description{
A \link[mlr3:LearnerRegr]{mlr3::LearnerRegr} for a regression k-Nearest-Neighbor implemented in \code{\link[kknn:kknn]{kknn::kknn()}} in package \CRANpkg{kknn}.
}
\keyword{datasets}
