% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_regr_xgboost.R
\docType{data}
\name{mlr_learners_regr.xgboost}
\alias{mlr_learners_regr.xgboost}
\alias{LearnerRegrXgboost}
\title{Regression xgboost Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrXgboost
}
\description{
An eXtreme Gradient Boosting learner for regression, implemented in \code{\link[xgboost:xgb.train]{xgboost::xgb.train()}}.
}
\keyword{datasets}
